bisher nur check if csv ergänzt

+++ src/DatabaseFile.py
@@ -45,19 +45,22 @@ class DatabaseFile:
         return pd.read_sql_query(self.cursor.execute(command), self.connection)
 
     def saveFile(self, file, name, seperator):
-        # TODO check if csv
-        file.save("name.csv")
-        if seperator == ',':
-            pd.read_csv("name.csv", sep=',').to_sql(name, sql.connect("Datenbank/file", check_same_thread=False),
-                                                schema=None, if_exists='replace', index=True, index_label=None,
-                                                chunksize=None,
-                                                dtype=None, method=None)
-        elif seperator == ';':
-            pd.read_csv("name.csv", sep=';').to_sql(name, sql.connect("Datenbank/file", check_same_thread=False),
-                                                schema=None, if_exists='replace', index=True, index_label=None,
-                                                chunksize=None,
-                                                dtype=None, method=None)
-        os.remove("name.csv")
+        filename = file.filename
+        namesplitted = filename.split('.')
+        last = namesplitted.pop()
+        if last == 'csv':
+            file.save("name.csv")
+            if seperator == ',':
+                pd.read_csv("name.csv", sep=',').to_sql(name, sql.connect("Datenbank/file", check_same_thread=False),
+                                                    schema=None, if_exists='replace', index=True, index_label=None,
+                                                    chunksize=None,
+                                                    dtype=None, method=None)
+            elif seperator == ';':
+                pd.read_csv("name.csv", sep=';').to_sql(name, sql.connect("Datenbank/file", check_same_thread=False),
+                                                    schema=None, if_exists='replace', index=True, index_label=None,
+                                                    chunksize=None,
+                                                    dtype=None, method=None)
+            os.remove("name.csv")
 
 
     def saveDataFrame(self, file, name):
